smal fix warp, warp speed display in item info, correct warp speed values in space for ships (Catalyst, Drake etc.). 30-07-2026

+++ src/eve-common/destiny/DestinyStructs.h
@@ -158,7 +158,7 @@ struct WARP_Struct {
     int32 effectStamp;   //statestamp of when warp started
     int64 followRange;   //unknown   -4616189618054758400 when warp is initiated.  calculation unknown for other values (used during warp)
     int64 followID;      //unknown   4669471951536783360 when warp is initiated or ship enters new bubble (AddBalls), 0 otherwise
-    int32 speed;
+    double speed;        // меняем на дубль - без округления
 };
 
 struct ORBIT_Struct {
